Suppress deletion of K8S objects during failover
The Volume Attachments and Pods should be deleted when they
disappear from the K8S data-store not during reconciliation
as this seems to get confused once a second instance of the
Pod is created (the creating one).

This can happen if reconnection to the Controller is delayed for
example due to it failing over.  In this case K8S detects the
lost node and begins creating the 'blocked' replacement Pod.